chore: resolve lint warnings and refresh documentation
- next.config.ts: extract resolveGithubUrl() to satisfy ESLint
  strict-boolean-expressions; correct stale workflow reference in comment
- contracts: suppress provably-safe forge-lint unsafe-typecast warnings in
  TrustLedger/JurorRegistry; drop redundant cast in JurorRegistry test
- docs: correct test counts (146 Hardhat, 84 Foundry, +4 fork), script
  tables, IPFS provider (Pinata, not Web3.Storage), and workflow list
  across README, src/README, and docs/*
- docs: fix markdownlint errors (fence spacing, line length, TOC anchors)
- lib: document decryptFile and hashString as complete-but-unused helpers
  retained for future use

+++ contracts/test/unit/JurorRegistryTest.t.sol
@@ -14,7 +14,7 @@ import {JurorRegistry} from "../../src/JurorRegistry.sol";
 contract JurorRegistryTest is Test {
     uint256 public constant MIN_STAKE = 0.01 ether;
     uint256 public constant LOCK_PERIOD = 7 days;
-    uint256 public constant COOLDOWN_PERIOD = 7 days; // mirrors JurorRegistry.JUROR_COOLDOWN
+    uint64 public constant COOLDOWN_PERIOD = 7 days; // mirrors JurorRegistry.JUROR_COOLDOWN
     uint256 public constant MIN_REPUTATION = 20; // mirrors JurorRegistry.MIN_REPUTATION
 
     JurorRegistry public registry;
@@ -297,7 +297,7 @@ contract JurorRegistryTest is Test {
 
         assertEq(
             registry.getCooldownUntil(juror1),
-            unlockTs + uint64(COOLDOWN_PERIOD),
+            unlockTs + COOLDOWN_PERIOD,
             "cooldown should be set to unlock time + JUROR_COOLDOWN"
         );
     }
